+++ R/cohort_performance_bins.R
@@ -1,0 +1,16 @@
+#' Cohort Performance Bins
+#'
+#' @param sirs301_obj a sirs301 object
+#' @param studentids students you want to plot
+#' @param subject c('ELA', 'Math')
+#'
+#' @return ggplot2 object
+#' @export
+
+cohort_performance_bins <- function(
+  sirs301_obj,
+  studentids,
+  subject
+) {
+
+}