+++ R/globals.R
@@ -1,0 +1,3 @@
+utils::globalVariables(
+  c("%>%", ".", "curr_grade_lvl", "discard", "discard1", "discard2", "end_year", "report_school_year", "start_year", "test_subject")
+)