+++ R/sirs301.R
@@ -1,17 +1,18 @@
 #' @title Create a sirs301 object
+#' @import ggplot2 magrittr
 #'
 #' @description
 #' \code{sirs301} creates a sirs301 object, enabling analysis and reporting.
 #'
 #' @param csvs data frame of sirs 301 exports - eg output of
 #' sirs301_import_csvs
+#' @param cohort_kind should we name cohorts by their college entry or college graduation year?
 #' @param verbose should sirs301 print status updates?  default is FALSE.
 #' @param ... additional arguments to pass to constructor functions
+#'
 #' @export
 
-sirs301 <- function(
-  csvs, cohort_kind = 'college_entry', verbose = FALSE, ...
-) UseMethod("sirs301")
+sirs301 <- function(csvs, cohort_kind = 'college_entry', verbose = FALSE, ...) UseMethod("sirs301")
 
 #' @export
 sirs301.default <- function(csvs, cohort_kind, verbose = FALSE, ...) {
@@ -25,6 +26,7 @@ sirs301.default <- function(csvs, cohort_kind, verbose = FALSE, ...) {
       curr_grade_lvl = ifelse(curr_grade_lvl == 'KF', 0, curr_grade_lvl),
       curr_grade_lvl = as.integer(curr_grade_lvl)
     )
+
 
   #get academic year
   df <- df %>%
@@ -49,12 +51,40 @@ sirs301.default <- function(csvs, cohort_kind, verbose = FALSE, ...) {
   nyseslat <- df %>%
     dplyr::filter(
       grepl('NYSESLAT', df$assessment_description, fixed = TRUE)
+    ) %>%
+    #break out assessment_description
+    tidyr::separate(
+      col = assessment_description,
+      into = c('test_subject', 'test_grade', 'discard1', 'discard2'),
+      sep = ' ',
+      remove = FALSE,
+      convert = TRUE
+    ) %>%
+    dplyr::select(
+      -discard1, -discard2
+    ) %>%
+    dplyr::mutate(
+      test_subject = gsub(':', '', test_subject, fixed = TRUE)
     )
 
   #grab science
   sci <- df %>%
     dplyr::filter(
       grepl('Sci: Scale', df$assessment_description, fixed = TRUE)
+    ) %>%
+    #break out assessment_description
+    tidyr::separate(
+      col = assessment_description,
+      into = c('discard1', 'test_grade', 'test_subject', 'discard2'),
+      sep = ' ',
+      remove = FALSE,
+      convert = TRUE
+    ) %>%
+    dplyr::select(
+      -discard1, -discard2
+    ) %>%
+    dplyr::mutate(
+      test_subject = gsub(':', '', test_subject, fixed = TRUE)
     )
 
   #grab ela / math
@@ -62,6 +92,17 @@ sirs301.default <- function(csvs, cohort_kind, verbose = FALSE, ...) {
     dplyr::filter(
       grepl(' ELA', df$assessment_description, fixed = TRUE) |
       grepl(' Math', df$assessment_description, fixed = TRUE)
+    ) %>%
+    #break out assessment_description
+    tidyr::separate(
+      col = assessment_description,
+      into = c('discard', 'test_grade', 'test_subject'),
+      sep = ' ',
+      remove = FALSE,
+      convert = TRUE
+    ) %>%
+    dplyr::select(
+      -discard
     )
 
   out <- list(