+++ NEWS.md
@@ -1,3 +1,8 @@
+# NYSEDtools 0.0.5 | 08/01/2016
+
+* `sirs301` object creation breaks assessment_description into `test_subject` and `test_grade`.
+
+
 # NYSEDtools 0.0.4 | 08/01/2016
 
 * `calculate_cohort` utility function.